pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lexiora"
version = "0.1.0"
description = "基于multiagent和Graphrag的法律咨询助手"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.40.0",
    "pyyaml>=6.0.2",
    "neo4j>=5.28.2",
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-core>=0.3.79",
    "langchain-community>=0.3.31",
    "langgraph>=1.0.8",
    "streamlit>=1.49.0",
    "dashscope>=1.20.0",
    "sse-starlette>=3.3.4",
    "fastapi>=0.136.0",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
namespaces = false
